+++ cmp-worker-kmp/src/jvmMain/kotlin/io/github/mobilebytelabs/worker/PlatformWorkManager.jvm.kt
@@ -9,7 +9,7 @@ actual object PlatformWorkManager {
                 "Add the worker-desktop module to jvmMain dependencies.",
         )
 
-    fun configure(workManager: WorkManager) {
+    fun configure(workManager: WorkManager?) {
         _delegate = workManager
     }
 }